Finialized integration of API into objects
