--- GERTe/GEDS/API.h
@@ -1,5 +1,0 @@
-#include "Gateway.h"
-#include "Peer.h"
-
-void processGateway(Gateway*);
-void processGEDS(Peer*);
